+++ baksmali/src/main/java/org/jf/baksmali/Adaptors/CommentMethodItem.java
@@ -5,16 +5,16 @@ import org.antlr.stringtemplate.StringTemplateGroup;
 
 public class CommentMethodItem extends MethodItem {
     private final StringTemplate template;
-    private final int sortOrder;
+    private final double sortOrder;
 
-    public CommentMethodItem(StringTemplateGroup stg, String comment, int codeAddress, int sortOrder) {
+    public CommentMethodItem(StringTemplateGroup stg, String comment, int codeAddress, double sortOrder) {
         super(codeAddress);
         template = stg.getInstanceOf("Comment");
         template.setAttribute("Comment", comment);
         this.sortOrder = sortOrder;
     }
 
-    public int getSortOrder() {
+    public double getSortOrder() {
         return sortOrder;
     }
